backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
bleach==6.2.0
celery==5.5.3
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.6
django-cors-headers==4.8.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
django-filter==24.3
daphne==4.1.2
django-redis
mysqlclient==2.1.0